+++ SwissArmyTransformer/model/official/__init__.py
@@ -5,3 +5,4 @@ from .vit_model import ViTModel
 from .bert_model import BertModel
 from .roberta_model import RobertaModel
 from .mae_model import MAE
+from .distill_model import DistillModel
