+++ examples/bert/README.md
@@ -27,3 +27,14 @@ bash scripts/finetune_adapter_boolq.sh /data/qingsong/pretrain /data/qingsong/da
 ```
 
 * iter 1000 val acc: 0.636
+
+Finetune bert by distillation:
+
+* First, train a bert large with scripts/finetune_boolq.sh
+* Then, distill by the following command
+
+```bash
+bash scripts/finetune_distill_boolq.sh /data/qingsong/pretrain /data/qingsong/dataset checkpoints/finetune-bert-large-uncased-boolq06-21-05-18
+```
+
+You should replace `checkpoints/finetune-bert-large-uncased-boolq06-21-05-18` to your own teacher model.
